refactor: merge dynamics.lineshape into dynamics
This allows for cleaner syntax like:
from ampform.dynamics import relativistic_breit_wigner
